Refactor Linkerd integration into addons package
Moved Linkerd-related functions from `cluster` to a new `addons` package for better modularity and organization. This improves code structure and aligns with the broader separation of concerns in managing Kubernetes addons.
